har-router-system/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-multipart==0.0.9
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0
